requirements.txt
Django<3
# psycopg2-binary
djangorestframework


